refactor(brand-search-optimization): migrate recipe to contrib and align with guidelines
