+++ contrib/python/brand-search-optimization/.env.example
@@ -1,0 +1,11 @@
+# Google Cloud Configuration
+GOOGLE_CLOUD_PROJECT=<TODO: update-this-value>
+GOOGLE_CLOUD_LOCATION=us-central1
+
+# Model Configuration (Gemini 3.7 Flash with Computer Use)
+MODEL_NAME=gemini-3.7-flash
+
+# BigQuery Catalog Dataset & Table
+DATASET_ID=products_data_agent
+TABLE_ID=shoe_items
+
