+++ contrib/python/brand-search-optimization/brand_search_optimization/sub_agents/keyword_finding/agent.py
@@ -1,0 +1,32 @@
+# Copyright 2026 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Defines keyword finding agent."""
+
+import os
+
+from google.adk.agents.llm_agent import Agent
+
+from ...tools import bq_connector
+from . import prompt
+
+keyword_finding_agent = Agent(
+    model=os.getenv("MODEL_NAME"),
+    name="keyword_finding_agent",
+    description="A helpful agent to find keywords",
+    instruction=prompt.KEYWORD_FINDING_AGENT_PROMPT,
+    tools=[
+        bq_connector.get_product_details_for_brand,
+    ],
+)
