+++ contrib/python/brand-search-optimization/deployment/bq_data_setup.sql
