+++ contrib/python/brand-search-optimization/manifest.yaml
@@ -1,0 +1,36 @@
+type: "standalone"    # Options: [standalone | module]
+status: "active"      # Options: [active | inactive]
+large: true
+language: "python"    # Options: [python | java | go | kotlin | typescript]
+description: "Brand Search Optimization agent using Gemini Computer Use to analyze live search rankings and optimize retail product titles."
+
+architecture:
+  agent: "multi"       # Options: [single | multi]
+  stateful: false      # Options: [true | false]
+  datasources:         # Options: [hardcoded | local | external]
+    - "external"
+
+dependencies:
+  libraries:
+    - "ADK"
+    - "Playwright"
+    - "Pillow"
+    - "BigQuery"
+    - "Pydantic"
+  services:
+    - "Gemini Computer Use"
+    - "Vertex AI"
+    - "BigQuery"
+
+ownership:
+  team: "Google"
+  poc: "anaelledrai"
+
+tags:
+  - "multi-agent"
+  - "computer-use"
+  - "retail"
+  - "brand-search-optimization"
+  - "seo"
+
+license: "Apache-2.0"
