--- python/agents/brand-search-optimization/brand_search_optimization/sub_agents/comparison/agent.py
@@ -1,40 +1,0 @@
-# Copyright 2025 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from google.adk.agents.llm_agent import Agent
-
-from ...shared_libraries import constants
-from . import prompt
-
-comparison_generator_agent = Agent(
-    model=constants.MODEL,
-    name="comparison_generator_agent",
-    description="A helpful agent to generate comparison.",
-    instruction=prompt.COMPARISON_AGENT_PROMPT,
-)
-
-comparsion_critic_agent = Agent(
-    model=constants.MODEL,
-    name="comparison_critic_agent",
-    description="A helpful agent to critique comparison.",
-    instruction=prompt.COMPARISON_CRITIC_AGENT_PROMPT,
-)
-
-comparison_root_agent = Agent(
-    model=constants.MODEL,
-    name="comparison_root_agent",
-    description="A helpful agent to compare titles",
-    instruction=prompt.COMPARISON_ROOT_AGENT_PROMPT,
-    sub_agents=[comparison_generator_agent, comparsion_critic_agent],
-)
